Update link rename to navLinks constants
