+++ src/components/navigation/Navbar.js
@@ -2,7 +2,7 @@ import React, { useState } from "react"
 import { Link } from "gatsby"
 import { GiHamburgerMenu } from "react-icons/gi"
 import { FaShoppingCart } from "react-icons/fa"
-import navLinks from "../../constants/links"
+import navLinks from "../../constants/navLinks"
 
 import styles from "./styles/navbar.module.scss"
